refactor: extract session stores and prepare libs for publishing
Split session stores out of express-govuk-starter into standalone
packages and prepare all three for npm publishing under @hmcts-cft scope.

- Extract @hmcts-cft/express-session-redis (Redis session middleware)
- Extract @hmcts-cft/express-session-postgres (Postgres session store
  with auto table creation and background expiry cleanup)
- Rename @hmcts/express-govuk-starter to @hmcts-cft/express-govuk-starter
- Remove Vite build config from the library, inline into web app
- Drop vite/sass/connect-redis/express-session peer deps from starter
- Add publishConfig metadata (repository, homepage, bugs, license, files)
  matching @hmcts-cft/simple-router pattern
