+++ apps/web/src/app.ts
@@ -1,15 +1,8 @@
 import path from "node:path";
 import { fileURLToPath } from "node:url";
 import { getPropertiesVolumeSecrets, hc, healthcheck, monitoringMiddleware } from "@hmcts/cloud-native-platform";
-import {
-  configureCookieManager,
-  configureGovuk,
-  configureHelmet,
-  configureNonce,
-  errorHandler,
-  expressSessionRedis,
-  notFoundHandler
-} from "@hmcts/express-govuk-starter";
+import { configureCookieManager, configureGovuk, configureHelmet, configureNonce, errorHandler, notFoundHandler } from "@hmcts-cft/express-govuk-starter";
+import { expressSessionRedis } from "@hmcts-cft/express-session-redis";
 import { createSimpleRouter } from "@hmcts-cft/simple-router";
 import cookieParser from "cookie-parser";
 import type { Express } from "express";
