+++ libs/express-session-postgres/src/index.ts
@@ -1,0 +1,3 @@
+export type { ExpressSessionPostgresOptions } from "./postgres-session.js";
+export { expressSessionPostgres } from "./postgres-session.js";
+export { PostgresStore } from "./postgres-store.js";
