+++ tsconfig.json
@@ -8,7 +8,9 @@
       "@hmcts/postgres": ["./apps/postgres/src"],
       "@hmcts/postgres-prisma": ["./libs/postgres-prisma/src"],
       "@hmcts/web": ["./apps/web/src"],
-      "@hmcts/express-govuk-starter": ["./libs/express-govuk-starter/src"],
+      "@hmcts-cft/express-govuk-starter": ["./libs/express-govuk-starter/src"],
+      "@hmcts-cft/express-session-redis": ["./libs/express-session-redis/src"],
+      "@hmcts-cft/express-session-postgres": ["./libs/express-session-postgres/src"],
       "@hmcts/cloud-native-platform": ["./libs/cloud-native-platform/src"],
       "@hmcts-cft/simple-router": ["./libs/simple-router/src"],
       "@hmcts/onboarding": ["./libs/onboarding/src"]
